handled different OS and if codecov already exists

+++ coverage.sh
@@ -18,14 +18,36 @@ if [ $rc -ne 0 ]; then
   exit $rc
 fi
 
+
+########## codecov starts ########3
 # make sure to set CODECOV_TOKEN env variable before doing this
-# codecov -f coverage.out
-# or 
-# bash <(curl -s https://codecov.io/bash) 
-curl -Os https://uploader.codecov.io/latest/linux/codecov
-chmod +x codecov
-./codecov -t ${CODECOV_TOKEN} -f coverage.out
-rm codecov
+os=`uname -s`
+
+if ! command -v codecov &> /dev/null
+then
+    # codecov uploader does not exist. Most likely in a CI environment.
+    if [ $os == "Darwin" ]; then
+    url="https://uploader.codecov.io/latest/macos/codecov"
+    elif [ $os == "Linux" ]; then
+        url="https://uploader.codecov.io/latest/linux/codecov"
+    elif [ $os == "Windows" ]; then
+        url="https://uploader.codecov.io/latest/linux/codecov"
+    else
+        echo "Unknown OS"
+        continue
+    fi
+
+    curl -Os $url
+    chmod +x codecov
+
+    ./codecov -t ${CODECOV_TOKEN} -f coverage.out
+    rm codecov
+else
+    codecov -t ${CODECOV_TOKEN} -f coverage.out
+fi
+
+########## codecov ends ########3
+
 
 # Example setup https://github.com/lluuiissoo/go-testcoverage/blob/main/.github/workflows/ci.yml
